Added environment variables and fixed pixel type

+++ museozuccante/api/models.py
@@ -8,8 +8,8 @@ class Room(models.Model):
     title = models.CharField(max_length=300)
     floor = models.IntegerField()
     number = models.IntegerField(default=0)
-    pixel_x = models.IntegerField(null=True, blank=True)
-    pixel_y = models.IntegerField(null=True, blank=True)
+    pixel_x = models.DecimalField(decimal_places=5, max_digits=10, null=True, blank=True)
+    pixel_y = models.DecimalField(decimal_places=5, max_digits=10, null=True, blank=True)
 
     def __str__(self):
         return self.title
